agentcoreDigestor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcoreDigestor"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "aws-opentelemetry-distro >= 0.10.0",
    "bedrock-agentcore >= 1.0.3",
    "mcp >= 1.19.0",
    "pytest >= 7.0.0",
    "pytest-asyncio >= 0.21.0",
    "python-dotenv >= 1.2.1",
    "strands-agents >= 1.13.0",
    "strands-agents-tools >= 0.2.16",
    "boto3 >= 1.34.0",
    "pandas >= 2.0.0",
    "pyarrow >= 16.0.0"
]

# --------------------------------------------------------------------------
# AgentCore-specific configuration
# --------------------------------------------------------------------------

[tool.agentcore.tools]
analyze_schema_arn = "arn:aws:lambda:eu-central-1:151441048511:function:agentcore-digestor-lambda-analyze-schema-dev"
validate_data_arn = "arn:aws:lambda:eu-central-1:151441048511:function:agentcore-digestor-lambda-validate-data-dev"
schema_normalizer_arn = "arn:aws:lambda:eu-central-1:151441048511:function:agentcore-digestor-lambda-schema-normalizer-dev"
ctas_arn = "arn:aws:lambda:eu-central-1:151441048511:function:agentcore-digestor-lambda-iceberg-ctas-dev"
load_into_iceberg_arn = "arn:aws:lambda:eu-central-1:151441048511:function:agentcore-digestor-lambda-load-into-iceberg-dev"
